// src/types/resume.types.ts
import { Types } from "mongoose";

export interface IPersonalInfo {
  fullname: string;
  email: string;
  mobile: string;
  location: string;
  github?: string;
  linkedIn?: string;
  portfolio?: string;
}

export interface IWorkExperience {
  company: string;
  position: string;
  startDate: Date;
  endDate: Date;
  description: string;
}

export interface IProjects {
  title: string;
  description: string;
  gitLink: string;
  liveLink: string;
  techStack: string[];
}

export interface IEducation {
  institute: string;
  degree: string;
  startDate: Date;
  endDate: Date;
  location?: string;
  gpa?: string;
}

export interface IResume {
    _id?: string;
    user_id: Types.ObjectId;
    title: string;
    summary: string;
    personalInfo: IPersonalInfo;
    workExperience?: IWorkExperience[];
    projects: IProjects[];
    education: IEducation[];
    certifications?: string[];
    skills?: string[];
}